task_service/requirements.txt
Django==4.2.7
django-rq==2.8.1
djangorestframework==3.14.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
pytest-redis==3.0.2
redis==5.0.1
rq==1.15.1
channels==4.0.0
djoser==2.2.2
drf-yasg==1.21.7
gunicorn==20.1.0


